requirements.txt
requests
json
logging
pyyaml
